think we got it setup properly now - show more
Cannot get partial row updates working. Like right now if you mark a work tracker as accepted, then it basically takes the local work tracker row and sends an update with everything in the row. so if this change was made offline and the remote changed some other part of the work tracker, then this will silently be overridden with the clients last state.

on the plus side we now have more robust dependencies between the objects. because in our supaLegend folder we set it up in the config, then we import every store into the init file, and then import the init into the root so this actually starts up all the observers properly.

Also we figured out that the changesSince being set to "all" (overriding our global setting) is necessary for the dependent tables to update. for example, we load in a work tracker that has an address. we're using a compute to take the address ids from all the work trackers, and then put them into an array and then use this array to sync with the remote. then if the work tracker changes to reference a different address, this needs to update the address observable to re sync with the new ideas to the remote. this is only possible with changesSince" "all" for these tables.

Currentl have a lot of useless code here like hopefully the entire db directory could be deleted. to solely rely on legend state.